feat: Implemented search functionality and fixed some bugs.
