+++ app/src/main/java/com/indisparte/cinemates/ui/MainActivity.kt
@@ -41,7 +41,8 @@ class MainActivity : AppCompatActivity(), ToFlowNavigable {
         // menu should be considered as top level destinations.
          val appBarConfiguration = AppBarConfiguration(
              setOf(
-                 R.id.homeFragment
+                 R.id.homeFragment,
+                 R.id.searchFragment
              )
          )
