--- app/src/main/res/layout/fragment_search.xml
@@ -1,90 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <com.google.android.material.appbar.AppBarLayout
-        android:id="@+id/appBarLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:elevation="4dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <androidx.appcompat.widget.Toolbar
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            android:background="@color/lynch_100"
-            app:layout_scrollFlags="scroll|enterAlways"
-            app:menu="@menu/search_menu"
-            app:navigationIcon="@drawable/ic_outline_arrow_back_24"
-            app:popupTheme="@style/ThemeOverlay.AppCompat.Dark">
-
-            <androidx.appcompat.widget.SearchView
-                android:id="@+id/search_view"
-                android:layout_width="match_parent"
-                android:layout_height="?attr/actionBarSize"
-                android:focusable="true"
-                android:focusedByDefault="true"
-                android:imeOptions="actionSearch"
-                android:inputType="text"
-                app:iconifiedByDefault="false"
-                app:queryBackground="@null"
-                app:queryHint="Search something..."
-                app:searchIcon="@null" />
-
-        </androidx.appcompat.widget.Toolbar>
-
-    </com.google.android.material.appbar.AppBarLayout>
-
-    <com.google.android.material.chip.ChipGroup
-        android:id="@+id/chip_group"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:clipToPadding="false"
-        android:orientation="horizontal"
-        android:padding="5dp"
-        app:checkedChip="@id/movies_chip"
-        app:chipSpacingHorizontal="16dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/appBarLayout"
-        app:selectionRequired="true"
-        app:singleSelection="true">
-
-        <com.google.android.material.chip.Chip
-            android:id="@+id/movies_chip"
-            style="@style/CustomFilterChipStyle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Movies" />
-
-        <com.google.android.material.chip.Chip
-            android:id="@+id/actors_chip"
-            style="@style/CustomFilterChipStyle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Actors" />
-
-        <com.google.android.material.chip.Chip
-            android:id="@+id/tv_show_chip"
-            style="@style/CustomFilterChipStyle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="TV Show" />
-    </com.google.android.material.chip.ChipGroup>
-
-    <com.indisparte.ui.custom_view.EmptyStateRecyclerView
-        android:id="@+id/recycler_view"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        app:emptyStateTitle="Nothing here!"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/chip_group" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
