+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -4,19 +4,10 @@
         android:id="@id/home_graph"
         android:icon="@drawable/ic_baseline_house_siding_24"
         android:title="Home" />
-
-  <!--  <item
+ <!--   <item
         android:id="@+id/discoverFragment"
         android:icon="@drawable/ic_discover"
         android:title="Discover" />
-    <item
-        android:id="@+id/savedFragment"
-        android:icon="@drawable/bookmark_icon"
-        android:title="To See" />
-
-    <item
-        android:id="@+id/profileFragment"
-        android:icon="@drawable/ic_outline_person_outline_24"
-        android:title="Profile" />-->
+   -->
 
 </menu>
