+++ core/navigation/src/main/java/com/indisparte/navigation/NavigationFlow.kt
@@ -3,7 +3,8 @@ package com.indisparte.navigation
 /**
  * @author Antonio Di Nuzzo (Indisparte)
  */
-sealed class NavigationFlow(val msg:String?) {
+sealed class NavigationFlow(val msg: String?) {
     object HomeFlow : NavigationFlow(null)
     class MovieDetailsFlow(movieId: Int) : NavigationFlow(movieId.toString())
+    object SearchFlow : NavigationFlow(null)
 }
