+++ core/navigation/src/main/res/navigation/main_navigation.xml
@@ -6,11 +6,18 @@
 
     <include app:graph="@navigation/home_graph" />
     <include app:graph="@navigation/movie_details_graph" />
+    <include app:graph="@navigation/search_graph" />
 
     <action
         android:id="@+id/action_global_home_flow"
         app:destination="@id/home_graph"
         app:popUpTo="@id/nav_graph" />
+
+    <action
+        android:id="@+id/action_global_search_flow"
+        app:destination="@id/search_graph"
+        app:enterAnim="@anim/slide_in_from_right"
+        app:exitAnim="@anim/slide_out_to_right" />
 
     <action
         android:id="@+id/action_global_movie_details"
