+++ core/navigation/src/main/res/values/nav_ids.xml
@@ -2,4 +2,5 @@
 <resources>
     <item name="home_graph" type="id" />
     <item name="movie_details_graph" type="id" />
+    <item name="search_graph" type="id" />
 </resources>
