+++ data/movie/src/main/java/com/indisparte/movie/source/MovieDataSource.kt
@@ -3,7 +3,7 @@ package com.indisparte.movie.source
 import com.indisparte.movie.response.CastDTO
 import com.indisparte.movie.response.CollectionDetailsResponseDTO
 import com.indisparte.movie.response.CrewDTO
-import com.indisparte.movie.response.MovieDTO
+import com.indisparte.response.MovieDTO
 import com.indisparte.movie.response.MovieDetailsDTO
 import com.indisparte.movie.response.ReleaseDatesByCountryDTO
 import com.indisparte.network.GenericResponse
