+++ data/search/src/main/java/com/indisparte/media_search/repository/MovieSearchRepository.kt
@@ -1,0 +1,13 @@
+package com.indisparte.media_search.repository
+
+import com.indisparte.model.entity.Movie
+import com.indisparte.network.Result
+import kotlinx.coroutines.flow.Flow
+
+/**
+ * @author Antonio Di Nuzzo
+ */
+interface MovieSearchRepository {
+
+    suspend fun searchMovieByTitle(title:String): Flow<Result<List<Movie>>>
+}
