+++ feature/media_search/build.gradle
@@ -1,23 +1,20 @@
 plugins {
-    id 'com.android.application'
+    id 'com.android.library'
     id 'org.jetbrains.kotlin.android'
-    id 'androidx.navigation.safeargs.kotlin'
-    id 'com.google.dagger.hilt.android'
     id 'kotlin-kapt'
+    id 'com.google.dagger.hilt.android'
 }
 
-
-
 android {
+    namespace 'com.indisparte.media_search'
     compileSdk libs.versions.targetSdk.get().toInteger()
 
     defaultConfig {
-        applicationId "com.indisparte.cinemates"
         minSdk libs.versions.minSdk.get().toInteger()
         targetSdk libs.versions.targetSdk.get().toInteger()
-        versionCode 1
-        versionName "1.0"
+
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
     }
 
     buildTypes {
@@ -26,32 +23,28 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+    buildFeatures {
+        viewBinding true
+        dataBinding true
+    }
 
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
-
-    buildFeatures {
-        viewBinding true
-        dataBinding true
+    kotlinOptions {
+        jvmTarget = '1.8'
     }
-    packagingOptions {
-        resources {
-            excludes += ['META-INF/DEPENDENCIES', 'META-INF/LICENSE', 'META-INF/LICENSE.txt', 'META-INF/license.txt', 'META-INF/NOTICE', 'META-INF/NOTICE.txt', 'META-INF/notice.txt', 'META-INF/ASL2.0']
-        }
-    }
-    namespace 'com.indisparte.cinemates'
 }
 
 dependencies {
 
+    implementation project(':core:util')
     implementation project(':core:ui')
+    implementation project(':core:network')
+    implementation project(':core:model')
     implementation project(':core:navigation')
-    implementation project(':feature:home')
-    implementation project(':feature:movie_details')
-    implementation project(':feature:media_search')
-
+    implementation project(':data:search')
 
     implementation(libs.appcompat)
     implementation(libs.material)
@@ -75,20 +68,7 @@ dependencies {
     implementation(libs.dagger.core)
     kapt(libs.dagger.compiler)
 
-    // Retrofit
-    implementation(libs.bundles.retrofit)
-
-
-    // Glide
-    implementation(libs.glide.core)
-    kapt(libs.glide.compiler)
-
-    // Youtube
-    implementation files('libs/YouTubeAndroidPlayerApi.jar')
-    implementation "org.jetbrains.kotlin:kotlin-reflect:1.8.22"
-
     //Timber
     implementation(libs.timber)
-
 
 }
