+++ feature/media_search/src/main/AndroidManifest.xml
