+++ feature/media_search/src/main/res/drawable/ic_back.xml
