+++ feature/movie_details/src/main/java/com/indisparte/movie_details/fragments/CollectionPartsFragment.kt
@@ -50,7 +50,7 @@ class CollectionPartsFragment : BaseFragment<FragmentCollectionPartsBinding>() {
                     LOG.d("Collection details loaded! $collectionDetails")
                     binding.progress.hide()
                     binding.collection = collectionDetails
-                    adapter.submitList(collectionDetails?.parts)
+                    adapter.submitList(collectionDetails.parts)
 
                 },
                 onError = { error ->
