--- feature/search/build.gradle
@@ -1,74 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'org.jetbrains.kotlin.android'
-    id 'kotlin-kapt'
-    id 'com.google.dagger.hilt.android'
-}
-
-android {
-    namespace 'com.indisparte.search'
-    compileSdk libs.versions.targetSdk.get().toInteger()
-
-    defaultConfig {
-        minSdk libs.versions.minSdk.get().toInteger()
-        targetSdk libs.versions.targetSdk.get().toInteger()
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles "consumer-rules.pro"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-    buildFeatures {
-        viewBinding true
-        dataBinding true
-    }
-
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = '1.8'
-    }
-}
-
-dependencies {
-
-    implementation project(':core:util')
-    implementation project(':core:ui')
-    implementation project(':core:network')
-    implementation project(':core:model')
-    implementation project(':core:navigation')
-    implementation project(':data:search')
-
-    implementation(libs.appcompat)
-    implementation(libs.material)
-    implementation(libs.constraintLayout)
-    implementation(libs.legacy)
-    implementation(libs.fragmentKtx)
-    testImplementation(libs.junitTest)
-    androidTestImplementation(libs.junitAndroidTest)
-    androidTestImplementation(libs.espresso)
-
-    // Kotlin Extensions
-    implementation(libs.core.ktx)
-
-    // Kotlin Coroutines
-    implementation(libs.bundles.coroutines)
-
-    // Jetpack Navigation - Kotlin
-    implementation(libs.bundles.navigation)
-
-    // Dagger Hilt
-    implementation(libs.dagger.core)
-    kapt(libs.dagger.compiler)
-
-    //Timber
-    implementation(libs.timber)
-
-}
