--- feature/search/src/main/java/com/indisparte/search/fragments/SearchViewModel.kt
@@ -1,19 +1,0 @@
-package com.indisparte.search.fragments
-
-import androidx.lifecycle.ViewModel
-import com.indisparte.search.repository.MovieSearchRepository
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-
-/**
- * @author Antonio Di Nuzzo
- */
-@HiltViewModel
-class SearchViewModel
-@Inject
-constructor(
-    private val movieSearchRepository: MovieSearchRepository,
-) : ViewModel() {
-
-}
