+++ settings.gradle
@@ -28,4 +28,4 @@ include ':feature:home'
 include ':feature:movie_details'
 include ':core:navigation'
 include ':data:search'
-include ':feature:search'
+include ':feature:media_search'
